empty route for spreadsheet export.
