+++ .gitignore
@@ -90,6 +90,8 @@ ENV/
 env.bak/
 venv.bak/
 
+.flaskenv
+
 # Spyder project settings
 .spyderproject
 .spyproject
