+++ requirements-dev.txt
@@ -3,3 +3,4 @@ coverage
 nose==1.3.7
 pylint
 xlrd
+python-dotenv
