+++ requirements.txt
@@ -1,14 +1,10 @@
-certifi==2018.4.16
-chardet==3.0.4
 click==6.7
 et-xmlfile==1.0.1
 Flask-Cors==3.0.3
-idna==2.6
 itsdangerous==0.24
-jdcal==1.4
 MarkupSafe==1.1
 werkzeug>=0.15.3
--e git+https://github.com/ebi-ait/ingest-client.git@7f04ab9a#egg=hca_ingest
+-e git+https://github.com/ebi-ait/ingest-client.git@6fd08d7#egg=hca_ingest
 jsonpickle~=1.2
 expiringdict==1.1.4
 jsonpath-rw
@@ -18,5 +14,4 @@ PyYAML~=5.1
 mock~=3.0.5
 dataclasses~=0.6
 argparse~=1.4.0
-
 jsonref~=0.2
